rename testCreateRequest file and update references
